LayoutTests/fast/preloader/first-base-tag-scanned-wins.html

<body>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpResourceResponseMIMETypes();
}
</script>
This test requires DumpRenderTree to see the log of what resources are loaded.
<script src=resources/non-existant.js></script>
<script>document.write("<plaintext>");</script>
<base href="resources/foo.html">
<base href="bogus/foo.html">
<img src="base-image3.png">
