Safari-609.1.11.1/LayoutTests/fast/preloader/first-base-tag-scanned-wins-expected.txt

base-image3.png has MIME type image/png
This test requires DumpRenderTree to see the log of what resources are loaded.

<base href="resources/foo.html">
<base href="bogus/foo.html">
<img src="base-image3.png">
